+++ servers/api/models/dataset.js
@@ -2,6 +2,34 @@
 
 var helpers = require('../support/helpers');
 
+var panel = require('../support/panel');
+var models; // Set dynamically in the associate.
+
+function calculateStatistics(dataset, options, callback) {
+    panel.readPanelJSON(null, dataset.id, {
+        properties: {},
+        statistics: {}
+    }, function (err, result) {
+        if (err) {
+            callback(err);
+            return;
+        }
+
+        dataset.startTime = result.startTime;
+        dataset.endTime = result.endTime;
+        dataset.summaryStatistics = result.summaryStatistics;
+        dataset.boundingBox = result.boundingBox;
+        dataset.boundingCircle = result.boundingCircle;
+        dataset.gpsLock = result.gpsLock;
+        dataset.source = {
+            scad: result.source
+        };
+        dataset.save();
+
+        callback(null, dataset);
+    });
+}
+
 module.exports = function (sequelize, DataTypes) {
     var dataset = sequelize.define('dataset', {
         startTime: helpers.createTimeField(DataTypes, 'startTime'),
@@ -27,7 +55,9 @@ module.exports = function (sequelize, DataTypes) {
     }, {
         freezeTableName: true,
         classMethods: {
-            associate: function (models) {
+            calculateStatistics: calculateStatistics,
+            associate: function (models_) {
+                models = models_;
                 dataset.belongsTo(models.user, {
                     constraints: true,
                     foreignKey: {
@@ -43,6 +73,9 @@ module.exports = function (sequelize, DataTypes) {
                 dataset.hasMany(models.event, constraints);
                 dataset.hasMany(models.comment, constraints);
                 dataset.hasMany(models.video, constraints);
+            },
+            getAssociations: function () {
+                return ['user', 'event', 'comment', 'video'];
             }
         },
         getterMethods: {