+++ servers/api/models/video.js
@@ -25,6 +25,9 @@ module.exports = function (sequelize, DataTypes) {
                         allowNull: false
                     }
                 });
+            },
+            getAssociations: function () {
+                return ['video'];
             }
         },
         hooks: {