+++ servers/api/routes/extra/sport.js
@@ -29,6 +29,15 @@ var sports = [
     },
     {
         name: 'paddleboard'
+    },
+    {
+        name: 'swim'
+    },
+    {
+        name: 'sail'
+    },
+    {
+        name: 'experiment'
     }
 ];