+++ servers/api/routes/utilities/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+exports.pickSearch = require('./picksearch');
+exports.replyMetadata = require('./replymetadata');
+exports.notFoundError = require('./notfounderror');